+++ code/maths/prime-checking.cpp
@@ -3,7 +3,7 @@ bool prime( int n ){
         return true;
     if( n % 2 == 0 || n <= 1 )
         return false;
-    for( int i = 3; i * i <= n; i += 2 )
+    for( long long i = 3; i * i <= n; i += 2 )
             if( ( n % i ) == 0 )
                 return false;
     return true;
